requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
Flask-Migrate>=4.0.0
Flask-Admin>=1.6.0
python-decouple>=3.8
Pillow>=10.0.0
wtforms~=3.2.1
werkzeug~=3.1.5